1523 Count odd numbers in interval range
Initially just looped through the list and incremented a count var based
on a conditional. This actually timed out and failed in leetcode which
makes sense. The faster way to do this and meet minimum requirements of
the problem is to simply take the high number - low number and divide by
two. With this approach you must take into consideration when edge cases
will break this logic. This only occurs when the low or hi number is
odd. In those situations you need to add 1 to the answer. This assumes
you are using ints not floats and int division gives you a whole number.

In python3 ints are converted to floats when you divide using /. // will
allow you to divide using just ints without converting to floats.